requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
pydantic
pydantic[email]
python-dotenv
